both updated backend

+++ controller/teacher-controller.js
@@ -276,9 +276,17 @@ const updateRank = async (req, res, next) => {
   let examIdObj = new mongoose.Types.ObjectId(examId);
   let delData = null;
   try {
-    delData = await McqRank.deleteMany({ examId: examIdObj });
+    delData = await McqRank.find({ examId: examIdObj });
   } catch (err) {
     return res.status(500).json("1.Something went wrong.");
+  }
+  if (delData.length > 0) {
+    let deleteData = null;
+    try {
+      deleteData = await McqRank.deleteMany({ examId: examIdObj });
+    } catch (err) {
+      return res.status(500).json("2.Something went wrong.");
+    }
   }
   let ranks = null;
   try {
@@ -288,7 +296,7 @@ const updateRank = async (req, res, next) => {
         totalObtainedMarks: -1,
       });
   } catch (err) {
-    return res.status(500).json("2.Something went wrong.");
+    return res.status(500).json("3.Something went wrong.");
   }
   //console.log("ranks:", ranks);
   let dataLength = ranks.length;
@@ -306,7 +314,7 @@ const updateRank = async (req, res, next) => {
   try {
     sav = await McqRank.insertMany(dataIns, { ordered: false });
   } catch (err) {
-    return res.status(500).json("3.Something went wrong.");
+    return res.status(500).json("4.Something went wrong.");
   }
   return res.status(201).json("Success!");
 };
